pyproject.toml
[project]
name = "ppap"
version = "0.1.0"
description = "Privacy-Preserving AI Proxy"
authors = [
    {name = "Mamunur Akand mamunur.iut@gmail.com"}
]
license = {text = "BUSINESS SOURCE LICENSE 1.1"}
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "spacy (==3.4.4)",
    "openai (==0.27.0)",
    "python-dotenv (==0.21.0)",
    "requests (==2.28.1)",
    "thinc (==8.1.0)",
    "numpy (==1.24.3)",
    "presidio-analyzer (>=2.2.358,<3.0.0)",
    "presidio-anonymizer (>=2.2.358,<3.0.0)"
]

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[tool.poetry]
packages = [{ include = "app" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
